Bug 641718 - Convert to using microseconds for internal video timestamps. r=kinetik
UltraBlame original commit: 44d43f095a4f34ae8c4cb85bdb29baf9f6c9717d

+++ content/media/wave/nsWaveReader.h
@@ -77,12 +77,12 @@ private:
   
   
   
-  float BytesToTime(PRInt64 aBytes) const;
+  double BytesToTime(PRInt64 aBytes) const;
 
   
   
   
-  PRInt64 TimeToBytes(float aTime) const;
+  PRInt64 TimeToBytes(double aTime) const;
